refactor: improve `Trampoline`
